# re-add optimizations when querying sparse data